# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "POS Staging API",
    "description": "Backend foundation and mock screens for the POS project.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/tables/{table_id}/open-session": {
      "post": {
        "tags": [
          "tables"
        ],
        "summary": "Open Table Session",
        "operationId": "open_table_session_api_v1_tables__table_id__open_session_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Table Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableSessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tables": {
      "get": {
        "tags": [
          "tables"
        ],
        "summary": "Table Dashboard",
        "description": "Dashboard bàn: mọi bàn của outlet, kèm phiên OPEN hiện tại (nếu có).\nMọi role đã đăng nhập đều xem được (xem trạng thái bàn không bị hạn chế\ntrong Permission Matrix).",
        "operationId": "table_dashboard_api_v1_tables_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TableDashboardItem"
                  },
                  "type": "array",
                  "title": "Response Table Dashboard Api V1 Tables Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tables/{table_id}": {
      "get": {
        "tags": [
          "tables"
        ],
        "summary": "Table Detail",
        "description": "Chi tiết bàn: phiên OPEN hiện tại (nếu có) + toàn bộ order của phiên đó,\nkèm tên món/thành tiền. Bàn không có phiên mở -> current_session=null, orders=[].",
        "operationId": "table_detail_api_v1_tables__table_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Table Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tables/{table_id}/mark-clean": {
      "post": {
        "tags": [
          "tables"
        ],
        "summary": "Mark Table Clean",
        "description": "CLEANING -> AVAILABLE. Chỉ bàn đang CLEANING mới xác nhận dọn xong được.",
        "operationId": "mark_table_clean_api_v1_tables__table_id__mark_clean_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Table Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OpenSessionRequest": {
        "properties": {
          "guest_count": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Guest Count",
            "default": 1
          }
        },
        "type": "object",
        "title": "OpenSessionRequest"
      },
      "OrderDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "table_session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Table Session Id"
          },
          "table_id": {
            "type": "string",
            "format": "uuid",
            "title": "Table Id"
          },
          "table_code": {
            "type": "string",
            "title": "Table Code"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_version": {
            "type": "integer",
            "title": "Current Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemDetailOut"
            },
            "type": "array",
            "title": "Items",
            "default": []
          },
          "subtotal": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Subtotal"
          }
        },
        "type": "object",
        "required": [
          "id",
          "table_session_id",
          "table_id",
          "table_code",
          "status",
          "current_version",
          "created_at",
          "subtotal"
        ],
        "title": "OrderDetailOut",
        "description": "Chi tiết 1 order cho màn POS ordering / order summary / table detail.\n`subtotal` = tổng các món chưa bị VOIDED (chưa có thuế/giảm giá — Pha 2)."
      },
...
      "OrderItemDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "menu_item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Menu Item Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Unit Price"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "menu_item_name": {
            "type": "string",
            "title": "Menu Item Name"
          },
          "line_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Line Total"
          }
        },
        "type": "object",
        "required": [
          "id",
          "menu_item_id",
          "status",
          "quantity",
          "unit_price",
          "note",
          "menu_item_name",
          "line_total"
        ],
        "title": "OrderItemDetailOut",
        "description": "Dòng món kèm tên món và thành tiền — đủ để FE vẽ màn order mà không phải\ngọi thêm GET /menu."
      },
...
      "TableDashboardItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "seats": {
            "type": "integer",
            "title": "Seats"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_session_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Session Id"
          },
          "guest_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guest Count"
          },
          "opened_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opened At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "seats",
          "status"
        ],
        "title": "TableDashboardItem",
        "description": "1 dòng của dashboard bàn: trạng thái bàn + phiên đang mở (nếu có)."
      },
      "TableDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "seats": {
            "type": "integer",
            "title": "Seats"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_session": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TableSessionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/OrderDetailOut"
            },
            "type": "array",
            "title": "Orders",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "seats",
          "status"
        ],
        "title": "TableDetailOut",
        "description": "Chi tiết 1 bàn (Figma screen-05): phiên đang mở + mọi order của phiên đó."
      },
      "TableOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "seats": {
            "type": "integer",
            "title": "Seats"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "seats",
          "status"
        ],
        "title": "TableOut"
      },
      "TableSessionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "table_id": {
            "type": "string",
            "format": "uuid",
            "title": "Table Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "guest_count": {
            "type": "integer",
            "title": "Guest Count"
          },
          "opened_at": {
            "type": "string",
            "format": "date-time",
            "title": "Opened At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "table_id",
          "status",
          "guest_count",
          "opened_at"
        ],
        "title": "TableSessionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}